+++ lib/multi_git/tree_entry.rb
@@ -3,21 +3,11 @@ require 'multi_git/object'
 require 'multi_git/builder'
 require 'multi_git/walkable'
 module MultiGit
-
-  base = Class.new
-
-  # @!parse
-  #    class TreeEntry < TreeEntry::Base
-  #    end
-  TreeEntry = Class.new(base)
-
   # A tree entry is like a {MultiGit::Object} or a {MultiGit::Builder} but it 
   # also has knows it's parent tree.
   class TreeEntry
 
-    Base = superclass
-
-    class Base
+    module Base
 
       include Walkable
 
@@ -56,11 +46,20 @@ module MultiGit
                   end
       end
 
-    end
+      def ==(other)
+        return false unless other.respond_to?(:path) && other.respond_to?(:object) && other.respond_to?(:mode)
+        return (path == other.path) && (object == other.object) && (mode == other.mode)
+      end
 
-    class Builder < Base
+      def inspect
+        ['#<', self.class.name,'@',path,' ', object.inspect, '>'].join
+      end
+   end
+
+    class Builder# < Base
 
       include MultiGit::Builder
+      include Base
 
       # @return [MultiGit::Builder]
       attr :object
@@ -78,24 +77,20 @@ module MultiGit
         super( parent, name , make_inner(*args) )
         instance_eval(&block) if block
       end
-
     end
 
     # @return [MultiGit::Object]
     attr :object
 
     include MultiGit::Object
+    include Base
 
     extend Forwardable
 
-    delegate (MultiGit::Object.instance_methods - ::Object.instance_methods) => :object
+    delegate [:oid, :content, :to_io, :bytesize] => :object
 
     def to_builder
       self.class::Builder.new(parent, name, object)
-    end
-
-    def inspect
-      ['#<', self.class.name,' ',path,' ', oid, '>'].join
     end